Repeticao

+++ src/funcionalidades.py
@@ -1,5 +1,5 @@
 import os
-from tkinter import filedialog, messagebox
+from tkinter import filedialog, messagebox, Tk
 
 def info_arquivo():
     """
@@ -14,8 +14,9 @@ def info_arquivo():
     caminho = get_caminho()
     os.chdir(caminho)
     try:
-        with open("info.txt", "w") as arquivo:
-            arquivo.write(string)
+        if not os.path.exists("info.txt"):
+            with open("info.txt", "w") as arquivo:
+                arquivo.write(string)
     except:
         messagebox.showerror("Erro", "Siga-nos nas redes sociais")
         
@@ -25,11 +26,15 @@ def get_caminho():
     Função para pegar o caminho da pasta a ser organizada.
     
     """
+    root = Tk()
+    root.withdraw()
+
     caminho = filedialog.askdirectory()
+    root.destroy()
     
     return caminho
 
-def moverArq(arquivo, extensao):
+def moverArq(arquivo):
     """Essa função verifica a extensão do arquivo e move para a pasta correspondente.
 
     Args:
@@ -38,23 +43,25 @@ def moverArq(arquivo, extensao):
     """
     
     # Criação de Tuplas para armazenar as extensões de arquivos
-    extensao_audio = ('.mp3', '.wav', '.flac', '.aac', '.ogg', '.wma', '.m4a', '.aiff', '.alac', '.amr', '.mid', '.midi', '.mpa', '.ra', '.ram', '.rmi', '.cda', '.m3u', '.pls', '.wpl', '.m4p', '.m4b', '.m4r', '.m4v', '.3gp', '.3g2', '.avi', '.flv', '.m4v', '.mkv', '.mov', '.mp4', '.mpg', '.mpeg', '.rm', '.swf', '.vob', '.wmv', '.asf', '.asx', '.m3u8', '.m3u', '.pls', '.xspf')
-    extensao_video = ('.3gp', '.3g2', '.avi', '.flv', '.m4v', '.mkv', '.mov', '.mp4', '.mpg', '.mpeg', '.rm', '.swf', '.vob', '.wmv', '.asf', '.asx', '.m3u8', '.m3u', '.pls', '.xspf')
-    extensao_imagem = ('.jpg', '.jpeg', '.png', '.gif', '.bmp', '.tiff', '.psd', '.pdf', '.eps', '.ai', '.indd', '.raw', '.svg', '.cdr', '.ico', '.webp', '.heif', '.heic', '.bpg', '.jxr', '.wdp', '.hdp', '.jng', '.jif', '.jfif', '.jpe', '.jfi', '.jp2', '.j2k', '.jpf', '.jpx', '.jpm', '.mj2', '.j2c', '.jpc', '.j2p', '.jps', '.jxl', '.jls','.jxr', '.wdp', '.hdp', '.jng', '.jif', '.jfif', '.jpe', '.jfi', '.jp2', '.j2k', '.jpf', '.jpx', '.jpm', '.mj2', '.j2c', '.jpc', '.j2p', '.jps', '.jxl', '.jls')
-    extensao_doc = ('.txt', '.doc', '.docx', '.odt', '.pdf', '.rtf', '.tex', '.wpd', '.wps', '.xml', '.msg', '.odt', '.odp', '.ods', '.odg', '.odf', '.sxw', '.sxi', '.sxc', '.sxd', '.sxi', '.sxm', '.sxg', '.stw', '.stc', '.std', '.stc', '.stw', '.sxw', '.sxd', '.sxi', '.sxm', '.sxg', '.sxc', '.sxd')
+    extensao_imagem = ('.jpg', '.jpeg', '.JPG', '.JPEG', '.png', '.PNG', '.gif', '.GIF', '.bmp', '.BMP', '.tiff', '.TIFF', '.svg', '.SVG', '.webp', '.WEBP','.heic', '.HEIC', '.ico', '.ICO', '.jfif', '.JFIF', '.pjpeg', '.PJPEG', '.pjp', '.PJP', '.avif', '.AVIF')
+    extensao_video = ('.mp4','.MOV','.avi', '.mkv', '.mov', '.wmv', '.flv', '.webm', '.mpeg', '.mpg', '.m4v','.MP4','.MOV','.AVI', '.MKV', '.MOV', '.WMV', '.FLV', '.WEBM', '.MPEG', '.MPG', '.M4V', '.hevc', '.HEVC', '.3gp', '.3GP', '.3g2', '.3G2', '.asf', '.ASF', '.avchd', '.AVCHD', '.flv', '.FLV', '.m2ts', '.M2TS', '.mts', '.MTS', '.mxf', '.MXF', '.ogg', '.OGG', '.rm', '.RM', '.rmvb', '.RMVB', '.ts', '.TS', '.vob', '.VOB', '.webm', '.WEBM', '.wmv', '.WMV')
+    extensao_doc = ('.pdf','.PDF' ,'.doc', '.docx', '.txt', '.xls', '.xlsx', '.ppt', '.pptx', '.odt', '.ods', '.odp','.odg', '.odf', '.odc', '.odb', '.odm', '.ott', '.ots', '.otp', '.otg', '.otf', '.otc', '.otb', '.oth', '.docm', '.dot', '.dotx', '.dotm', '.docb', '.xlsb', '.xlsm', '.xltx', '.xltm', '.xlsx', '.xlsm', '.xlt', '.xlm', '.xlw', '.pptm', '.potx', '.potm', '.ppam', '.ppsx', '.ppsm', '.sldx', '.sldm', '.pot', '.pps', '.ppa', '.ppam', '.ppsx', '.ppsm', '.sldx', '.sldm', '.pot', '.pps', '.ppa', '.ppam', '.ppsx', '.ppsm', '.sldx', '.sldm', '.pot', '.pps', '.ppa', '.ppam', '.ppsx', '.ppsm', '.sldx', '.sldm', '.pot', '.pps', '.ppa', '.ppam', '.ppsx', '.ppsm', '.sldx', '.sldm', '.pot', '.pps', '.ppa', '.ppam', '.ppsx', '.ppsm', '.sldx', '.sldm', '.pot', '.pps', '.ppa', '.ppam', '.ppsx', '.ppsm', '.sldx', '.sldm', '.pot', '.pps', '.ppa', '.ppam', '.ppsx', '.ppsm', '.sldx', '.sldm', '.pot', '.pps', '.ppa', '.ppam', '.ppsx', '.ppsm', '.sldx', '.sldm', '.pot', '.pps', '.ppa', '.ppam', '.ppsx', '.ppsm', '.sldx', '.sldm', '.pot', '.pps', '.ppa', '.ppam', '.ppsx', '.ppsm', '.sldx', '.sldm', '.pot', '.pps', '.ppa', '.ppam', '.ppsx', '.ppsm', '.sldx', '.sldm', '.pot', '.pps', '.ppa', '.ppam')
+    extensao_exe = ('.exe', '.bat', '.sh', '.bin', '.cmd', '.com', '.msi', '.jar', '.py', '.pl')
+    extensao_audio = ('.mp3', '.wav', '.flac', '.aac', '.ogg', '.wma', '.m4a', '.aiff', '.alac')
     extensao_compactado = ('.zip', '.rar', '.7z', '.tar', '.gz', '.tgz', '.bz2', '.tbz2', '.tar.bz2', '.tar.gz', '.tar.xz', '.tar.zst', '.tar.lz', '.tar.lzma', '.tar.lz4', '.tar.lzo', '.tar.xz', '.tar.zst', '.tar.zstd', '.tar.z','iso')     
-    extensao_executavel = ('.exe', '.msi', '.apk', '.app', '.bat', '.bin', '.cgi', '.com', '.gadget', '.jar', '.wsf', '.appref-ms', '.pif', '.vb', '.ws', '.msc', '.scr', '.reg', '.sh', '.bash', '.csh', '.ksh', '.sh', '.zsh', '.cmd', '.ps1', '.psm1', '.psd1', '.ps1xml', '.psc1', '.pssc', '.msh', '.msh1', '.msh2', '.mshxml', '.msh1xml', '.msh2xml', '.scf', '.lnk', '.inf', '.url', '.cplic', '.cplic')
     extensao_script = ('.sh', '.bash', '.csh', '.ksh', '.sh', '.zsh', '.cmd', '.ps1', '.psm1', '.psd1', '.ps1xml', '.psc1', '.pssc', '.msh', '.msh1', '.msh2', '.mshxml', '.msh1xml', '.msh2xml')        
     
     # Criação de dicionários para armazenar as extensão
-    arq = {"Audios": extensao_audio, "Videos": extensao_video, "Imagens": extensao_imagem, "Documentos": extensao_doc, "Compactados": extensao_compactado, "Programas": extensao_executavel, "Script": extensao_script}
+    arq = {"Audios": extensao_audio, "Videos": extensao_video, "Imagens": extensao_imagem, "Documentos": extensao_doc, "Compactados": extensao_compactado, "Programas": extensao_exe, "Script": extensao_script}
     
     try:
         # Verifica a extensão do arquivo e move para a pasta correspondente
         #percorrendo o dicionário
         for chave, valor in arq.items():
             #verificando se a extensão do arquivo está no valor(valor é uma tupla que contém as extensões)
-            if extensao in valor:
+            
+            #if extensao in valor:
+            if arquivo.endswith(valor):
                 #verificando se a pasta existe
                 if not os.path.exists(chave):
                     #Se não existir, cria a pasta
@@ -63,7 +70,7 @@ def moverArq(arquivo, extensao):
                 os.rename(arquivo, chave + '/' + arquivo)
                 break
     except:
-        messagebox.showerror("Erro", "Erro ao mover arquivo")   
+        messagebox.showerror("Erro", "Erro ao mover arquivo ")   
              
 def moverDoc():
     """Essa função verifica a extensão do arquivo e move para a pasta correspondente,
@@ -134,14 +141,11 @@ def organizar():
             # Verifica se o arquivo é um arquivo
             if os.path.isfile(arquivo):
                 
-                # Separa a extensão do arquivo
-                extensao = os.path.splitext(arquivo)[1]
-                
                 # Verifica a extensão do arquivo e move para a pasta correspondente
-                moverArq(arquivo, extensao)
+                moverArq(arquivo)
         
         # Organiza os documentos
-        moverDoc("Documentos")
+        #moverDoc("Documentos")
         
         # Gera informações das redes sociais em arquivo txt
         info_arquivo()
